Fixed some issues with java8 and generic to primitive types

+++ plugins/de.cau.cs.kieler.core.kgraph.text.ui/src/de/cau/cs/kieler/core/kgraph/text/ui/random/wizard/RandomGraphWizard.java
@@ -283,7 +283,7 @@ public class RandomGraphWizard extends Wizard implements INewWizard {
         if (options.getProperty(GeneratorOptions.TIME_BASED_RANDOMIZATION)) {
             random = new Random();
         } else {
-            random = new Random(options.getProperty(GeneratorOptions.RANDOMIZATION_SEED));
+            random = new Random(options.getProperty(GeneratorOptions.RANDOMIZATION_SEED).intValue());
         }
         RandomGraphGenerator generator = new RandomGraphGenerator(random);
